refactor: move Game component to its own file
